bake in debouncing to ColorPillPicker
